fix: fixing references from merge
